// src/app/store/profile/profile.actions.types.ts
export enum ProfileActionTypes {
  FETCH_PROFILE = '[PROFILE] Fetch Profile',
  FETCH_PROFILE_SUCCESS = '[PROFILE] Fetch Profile Success',
  FETCH_PROFILE_ERROR = '[PROFILE] Fetch Profile Error',
  UPDATE_PROFILE = '[PROFILE] Update Profile',
  UPDATE_PROFILE_SUCCESS = '[PROFILE] Update Profile Success',
  UPDATE_PROFILE_ERROR = '[PROFILE] Update Profile Error',
  DELETE_PROFILE = '[PROFILE] Delete Profile',
  DELETE_PROFILE_SUCCESS = '[PROFILE] Delete Profile Success',
  DELETE_PROFILE_ERROR = '[PROFILE] Delete Profile Error'
}